requirements.txt
glfw==2.10.0
numpy==2.3.5
pillow==12.1.0
pygame==2.6.1
pyglm==2.8.3
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
readchar==4.2.1
